Name Variable Updates
Essentially, after going crazy i realized, that just like with any program, you can't have the same names for variables, which was limiting the program from taking in diferent body messages. So I added 3, begin is for being, zipcode is for the zip, and demand will be for the command which determines the information the select.