Just reformated some stuff.
I moved the functions up to the top of the code as per the stylization techniques discussed in class. I commented out some code just in case we need it later, and/or it was repeated. I also moved some code into an existing loop just to simplify the code.