Full Functioning Program.
Well we ended up just parsing through and having strict parameters. We will need to do some controlling for spaces and upper cases assuming that works with the len() function. All thats left if changed the output to it looks good/ not as content heavy. And a few input updates.